Complete passTurn and turnStat
The passTurn Javascript macro was not taking variable references
correctly, so the same functionality was implemented within a
twee passage. Similarly, the turnStat() function was added to the
Mod Javascript class. Because of the shift, the "attack" passage
was moved to the same file as passTurn, and the file itself was
renamed to reflect its focus on mechanics.